feat: /books/{id} validate ID param

+++ routes/books.js
@@ -5,13 +5,25 @@ import { BooksRepository } from '../repositories/booksRepository.js';
 import { BooksService } from '../services/booksService.js';
 import { CreateBookDTO } from '../dtos/createBookDTO.js';
 import { UpdateBookDTO } from '../dtos/updateBookDTO.js';
+import { BookIdDTO } from '../dtos/bookIdDTO.js';
 
 const repository = new BooksRepository();
 const service = new BooksService(repository);
 
+const validateIdParam = (req, res, next) => {
+  const id = +req.params['id'];
+  if (BookIdDTO.validate(id)) {
+    return res.status(400).json({
+      message: 'The ID param is invalid'
+    });
+  }
+
+  next();
+}
+
 // Find one
-router.get('/:id', (req, res) => {
-  const id = req.params['id'];
+router.get('/:id', validateIdParam, (req, res) => {
+  const id = +req.params['id'];
   return res.json(service.listOne(id));
 });
 
@@ -37,8 +49,14 @@ router.post('/', (req, res) => {
 });
 
 // Delete
-router.delete('/:id', (req, res) => {
-  const id = req.params['id'];
+router.delete('/:id', validateIdParam, (req, res) => {
+  const id = +req.params['id'];
+  if (BookIdDTO.validate(id)) {
+    return res.status(400).json({
+      message: 'The ID param is invalid'
+    });
+  }
+
   try {
     service.delete(id);
     return res.status(204).send();
@@ -48,7 +66,7 @@ router.delete('/:id', (req, res) => {
 });
 
 // Update
-router.patch('/:id', (req, res) => {
+router.patch('/:id', validateIdParam, (req, res) => {
   const id = req.params['id'];
   const book = req.body;
   const hasKeys = UpdateBookDTO.validate(book);
